requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
asgiref==3.6.0
openai==0.27.5
revChatGPT==5.0.0